requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
imbalanced-learn>=0.12.0
xgboost>=2.0.0
tqdm>=4.66.0
rdkit-pypi>=2022.9.5
PyYAML>=6.0
